Renamed the model-index color theme file to trajectory-index
